refactor(contact-api): use route handlers instead of api routes + remove axios
