+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "axios": "^0.21.1",
     "formik": "^2.2.9",
     "next": "^13.4.4",
     "next-pwa": "^5.4.1",
